클래스 수정
